+++ apps/design-system/src/subjects/views/secrets/types.ts
@@ -1,0 +1,7 @@
+export enum ScopeEnum {
+  ACCOUNT = 'account',
+  ORGANIZATION = 'organization',
+  PROJECT = 'project'
+}
+
+export type SecretScope = 'account' | 'organization' | 'project'
