+++ packages/ui/src/views/secrets/secret-reference/secret-reference.tsx
@@ -1,4 +1,4 @@
-import { Button, ButtonGroup, Icon, StackedList } from '@/components'
+import { Alert, Button, ButtonGroup, Icon, StackedList } from '@/components'
 import { EntityReference, EntityRendererProps } from '@views/platform'
 import { DirectionEnum } from '@views/platform/types'
 
@@ -9,6 +9,7 @@ export interface SecretReferenceProps {
   secretsData: SecretItem[]
   childFolder: string | null
   parentFolder: string | null
+  apiError?: string | null
 
   // State
   selectedEntity: SecretItem | null
@@ -26,6 +27,7 @@ export const SecretReference: React.FC<SecretReferenceProps> = ({
   secretsData,
   childFolder,
   parentFolder,
+  apiError,
 
   // State
   selectedEntity,
@@ -51,8 +53,7 @@ export const SecretReference: React.FC<SecretReferenceProps> = ({
           <Button
             variant="default"
             size="sm"
-            onClick={e => {
-              e.stopPropagation()
+            onClick={() => {
               onSelect(entity)
             }}
           >
@@ -66,7 +67,7 @@ export const SecretReference: React.FC<SecretReferenceProps> = ({
   }
 
   return (
-    <div className="flex flex-col h-full">
+    <div className="flex flex-col">
       <span className="font-medium mb-4">Select an existing Secret:</span>
       <div className="flex-1">
         <EntityReference<SecretItem>
@@ -79,7 +80,13 @@ export const SecretReference: React.FC<SecretReferenceProps> = ({
           childFolder={childFolder}
           isLoading={isLoading}
         />
+        {apiError ? (
+          <Alert.Container variant="destructive" className="mt-4">
+            <Alert.Description>{apiError}</Alert.Description>
+          </Alert.Container>
+        ) : null}
       </div>
+
       <div className="bg-background-2 fixed bottom-0 left-0 right-0 p-4 shadow-md">
         <ButtonGroup className="flex flex-row justify-between">
           <Button type="button" variant="outline" onClick={onCancel}>
